refactor: services/metrics
