+++ services/agents/app/core/crews/ml_engeneer/ml_engeneer.py
@@ -8,7 +8,7 @@ from crewai.tools import tool
 
 from .tools import get_tools
 from ..utils import track_agent, get_agent_role_from_config
-from app.services.metrics.post import add_agent_in_metrics
+from app.services.metrics import add_agent_in_metrics
 from app.services.agent_history import agent_history_client
 from app.logs import get_logger
 from app.core.llm import llm
