--- services/agents/app/services/metrics/utils.py
@@ -1,36 +1,0 @@
-import requests
-from functools import wraps
-
-from app.config import config_url
-from app.logs import get_logger
-
-logger = get_logger(__name__)
-
-def get_json(endpoint: str) -> dict:
-    resp = requests.get(f"{config_url.METRICS_URL}{endpoint}")
-    resp.raise_for_status()
-    return resp.json()
-
-def handle_errors(func):
-    """Декоратор для обработки ошибок API запросов"""
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        try:
-            return func(*args, **kwargs)
-        except requests.exceptions.Timeout:
-            error_str = f"Timeout: {func.__name__} не ответил..."
-            logger.error(error_str)
-            return {"ERROR": error_str}
-        except requests.exceptions.ConnectionError:
-            error_str = f"Connection error: Не удалось подключиться к {config_url.METRICS_URL}"
-            logger.error(error_str)
-            return {"ERROR": error_str}
-        except requests.exceptions.HTTPError as e:
-            error_str = f"HTTP error {e.response.status_code}: {e.response.text}"
-            logger.error(error_str)
-            return {"ERROR": error_str}
-        except Exception as e:
-            error_str = f"Ошибка в {func.__name__}: {str(e)}"
-            logger.error(error_str)
-            return {"ERROR": error_str}
-    return wrapper
